Refactor function calling
Fixed repeated last VRMA animation
